requirements.txt
asgiref==3.2.3
bson==0.5.8
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dataclasses==0.6
Django==2.2.11
django-mongoengine==0.4.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djongo==1.3.1
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.8
Werkzeug==1.0.0
